Removing more ToDos and adding support for ServiceContexts

+++ src/main/java/BESA/SocialRobot/ServiceProvider/agent/ServiceProviderState.java
@@ -5,25 +5,25 @@ import BESA.Kernel.Social.ServiceProvider.agent.StateServiceProvider;
 
 public class ServiceProviderState extends StateServiceProvider {
 
-    private ServiceRequestManager<Integer> manager;
+    private ServiceRequestManager<Double> manager;
 
     public ServiceProviderState(ServiceProviderDescriptor descriptor) {
         super(descriptor);
         manager = new ServiceRequestManager<>();
     }
 
-    public ServiceRequestManager<Integer>  getRequestManager() {
+    public ServiceRequestManager<Double>  getRequestManager() {
         return manager;
     }
 
-    public boolean hasRequestPending(int id) {
+    public boolean hasRequestPending(double id) {
         return manager.hasRequestPending(id);
     }
-    public void addRequest(int id) {
+    public void addRequest(double id) {
         manager.addRequest(id);
     }
 
-    public void requestCompleted(int id) {
+    public void requestCompleted(double id) {
         manager.requestCompleted(id);
     }
 }
